Add material management module
- Adds module controller for component lifecycle handling
- Adds facade for external module access based on the `IMaterialManagement` facade interface
- Adds the `IContainerPool` component to monitor the known `IMaterialContainer` instances in the system
  - Register added, removed resources
  - Forward update/state-change events from containers
- Adds `IMaterialFlowHandler` component to handle material flow transitions
  - Adds material request handling
  - Adds material pre-advice handling
- Adds Integration tests for
  - retrieving material container types
  - material request handling
  - material pre-advice handling

Also
- Extends resources REST API to retrieve single resource type nodes

+++ src/Tests/Moryx.Material.IntegrationTests/TestContainerHost.cs
@@ -20,7 +20,7 @@ public class TestContainerHost : Resource
     /// Should create a <see cref="BasicMaterialContainer"/> in <see cref="States.StateClassification.Available"/>
     /// using the resource constructor for material registration.
     /// </summary>
-    public async Task<MaterialContainer> RegisterMaterialAsync(string material, decimal quantity, string? unit, CancellationToken cancellationToken = default)
+    public async Task<MaterialContainer> RegisterMaterialAsync(string material, double quantity, string? unit, CancellationToken cancellationToken = default)
     {
         var container = Graph.Instantiate<BasicMaterialContainer>().Configure(c => c.With(
             material: material,
